Refactor fromJson with XContent helper
